Create majorityelement.py
To find the majority element in an array of integers, you can use the Boyer-Moore Majority Vote algorithm. This algorithm efficiently finds the majority element in linear time complexity, without requiring any extra space.

Here's the implementation in Python to find the majority element in the given input [2, 4, 5, 5, 5, 5, 5]:

In the given input, the majority element is 5 because it appears more than n/2 times in the array, where n is the size of the array.