Create findDuplicates.java
The 'findDuplicate' method takes an array of integers as input and returns a list of elements that appear twice in the array. It uses a HashSet to keep track of elements encountered so far. If an element is already present in the set, it is added to the duplicates list. Otherwise, the element is added to the set. Finally, the main method demonstrates the usage of this method by finding duplicates in the given array and printing the result.